:recycle: Convert use-travel-authorization-pre-approval-profiles to TypeScript.
